add services and finish the functions that have relation with n_queens

+++ src/interfaces/individual.py
@@ -19,7 +19,7 @@ class Individual(ABC):
         pass
 
     @abstractmethod
-    def to_rate(self) -> float:
+    def to_rate(self) -> int:
         pass
 
     @property
